requirements.txt
langchain
langchain_groq
faiss-cpu
pandas
python-dotenv
fastapi
uvicorn
pydantic
langchain_community
huggingface-hub
beautifulsoup4
streamlit